Hatches.cpp Improving of zero round animation Change floating point calculations to integer
AI.cpp
Changes bots brain and identity float to integer

GameLogic.cpp
delete clear_chars()

Also:
Unpin currenthatch from F

git-svn-id: https://russian-roulette.googlecode.com/svn@33 1768935b-6245-56be-5c17-994056f11f20

+++ branches/Refactor/AI.h
@@ -9,8 +9,8 @@ enum TBotType {
 class TBot { // класс ботов
 
 private:
-	float bBrain;
-	float bIdentity;
+	int bBrain;
+	int bIdentity;
 
 public:
 	bool Get_Answer();
